@@ -2,7 +2,7 @@ package me.robwilliams.pack.fragment;
 
 import android.content.ContentValues;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.View;
